nlp_module/requirements.txt
openai>=1.0.0,<2.0.0
python-telegram-bot>=13.0,<21.0
